First commit regarding the integration of Mesh Header into the 6LoWPAN layer

+++ core/net/sdn/sdn.h
@@ -25,7 +25,7 @@ typedef enum {
  * \return An integer equals to -1 if some error occurs or if the parameter is 
  * not valid
  */
-uint8_t forward(uint8_t nextLayer);
+int forward(uint8_t nextLayer);
 
 #endif /* SDN_H_ */
